updated preprocessors to work with GPU

+++ preprocessors/meanpreprocessor.py
@@ -1,5 +1,5 @@
 # import the necessary packages
-import numpy as np
+import tensorflow as tf
 
 
 class MeanPreprocessor:
@@ -17,7 +17,7 @@ class MeanPreprocessor:
         determines if the input data is to be normalized to the range [0, 1]
     """
 
-    def __init__(self, mean, std, normalize = True):
+    def __init__(self, mean, std, normalize=True):
         # initialize the instance variables
         self.mean = mean
         self.std = std
@@ -26,7 +26,7 @@ class MeanPreprocessor:
     def preprocess(self, img):
         # check if the image is to be normalized
         if self.normalize:
-            img = img.astype(np.float32) / 255.0
+            img = tf.cast(img, tf.float32) / 255.0
 
         # return the processed image
         return (img - self.mean) / self.std
